fixup! feat: add endpoint to get listening ports in agent
