+++ agent/statsendpoint.go
@@ -2,13 +2,10 @@ package agent
 
 import (
 	"net/http"
-	"runtime"
 	"sync"
 	"time"
 
-	"github.com/cakturk/go-netstat/netstat"
 	"github.com/go-chi/chi"
-	"golang.org/x/xerrors"
 
 	"github.com/coder/coder/coderd/httpapi"
 	"github.com/coder/coder/codersdk"
@@ -32,53 +29,6 @@ type listeningPortsHandler struct {
 	mut   sync.Mutex
 	ports []codersdk.ListeningPort
 	mtime time.Time
-}
-
-func (lp *listeningPortsHandler) getListeningPorts() ([]codersdk.ListeningPort, error) {
-	lp.mut.Lock()
-	defer lp.mut.Unlock()
-
-	if runtime.GOOS != "linux" && runtime.GOOS != "windows" {
-		// Can't scan for ports on non-linux or non-windows systems at the
-		// moment. The UI will not show any "no ports found" message to the
-		// user, so the user won't suspect a thing.
-		return []codersdk.ListeningPort{}, nil
-	}
-
-	if time.Since(lp.mtime) < time.Second {
-		// copy
-		ports := make([]codersdk.ListeningPort, len(lp.ports))
-		copy(ports, lp.ports)
-		return ports, nil
-	}
-
-	tabs, err := netstat.TCPSocks(func(s *netstat.SockTabEntry) bool {
-		return s.State == netstat.Listen
-	})
-	if err != nil {
-		return nil, xerrors.Errorf("scan listening ports: %w", err)
-	}
-
-	ports := []codersdk.ListeningPort{}
-	for _, tab := range tabs {
-		if tab.LocalAddr.Port < uint16(codersdk.MinimumListeningPort) {
-			continue
-		}
-
-		ports = append(ports, codersdk.ListeningPort{
-			ProcessName: tab.Process.Name,
-			Network:     codersdk.ListeningPortNetworkTCP,
-			Port:        tab.LocalAddr.Port,
-		})
-	}
-
-	lp.ports = ports
-	lp.mtime = time.Now()
-
-	// copy
-	ports = make([]codersdk.ListeningPort, len(lp.ports))
-	copy(ports, lp.ports)
-	return ports, nil
 }
 
 // handler returns a list of listening ports. This is tested by coderd's
